Moving runTask to functional equivalent
