+++ addon/index.js
@@ -1,0 +1,2 @@
+export { runTask } from 'utils/run';
+export { runDisposables } from 'utils/disposable';
