+++ tests/unit/mixins/run-test.js
@@ -7,7 +7,6 @@ import ContextBoundTasksMixin, {
   setShouldPoll,
   pollTaskFor,
   cancelBoundTasks,
-  getTask,
 } from 'ember-lifeline/mixins/run';
 
 module('ember-lifeline/mixins/run', {
@@ -616,32 +615,4 @@ test('cancelBoundTasks cancel function is called once for each task', function(a
   cancelBoundTasks(tasks, cancelFn);
 
   assert.equal(callCount, 3, 'The cancel function is called 3 times');
-});
-
-test('getTask returns passed in task function as task', function(assert) {
-  assert.expect(1);
-
-  let task = () => {};
-
-  assert.equal(task, getTask(null, task, 'foo'), 'tasks are equal');
-});
-
-test('getTask returns passed in task from the instance', function(assert) {
-  assert.expect(1);
-
-  let instance = { fooTask: () => {} };
-
-  assert.equal(
-    instance.fooTask,
-    getTask(instance, 'fooTask', 'foo'),
-    'tasks are equal'
-  );
-});
-
-test('getTask throws when task not found', function(assert) {
-  assert.expect(1);
-
-  assert.throws(() => {
-    getTask({}, null, 'foo');
-  }, /You must pass a task function or method name to 'foo'./);
 });
